+++ NitroComposer/SequenceCommands/Rand/NoteRandCommand.cs
@@ -1,10 +1,10 @@
 ﻿namespace Nitro.Composer.SequenceCommands.Rand {
 	public class NoteRandCommand : BaseSequenceCommand {
-		public uint Note;
+		public byte Note;
 		public uint Velocity;
 		public short DurationMin, DurationMax;
 
-		public NoteRandCommand(uint note, uint vel, short durationMin, short durationMax) {
+		public NoteRandCommand(byte note, uint vel, short durationMin, short durationMax) {
 			Note = note;
 			Velocity = vel;
 			DurationMin = durationMin;
